requirements.txt
Authlib==1.5.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dnspython==2.7.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.26.1
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
uuid==1.30
Werkzeug==3.1.3